Reduce no. of file of history related code
