+++ internal/store/history_impl.go
@@ -24,3 +24,38 @@ func InsertPublisher(db *lazydb.LazyDB, value ...string) error {
 func GetPublisherList(db *lazydb.LazyDB) ([]string, error) {
 	return getHistory(db, definitions.CategoryPublisher)
 }
+
+// ================================================================================
+
+// Container for history module,
+// which wraps all necessary value will be used to insert value.
+//
+// This type is designed for insert value with different category at a time.
+type HistoryVal struct {
+	Category definitions.CategoryType // category to be inserted
+	Value    string                   // value to be inserted
+}
+
+// Insert values to database, which can include multiple record with different categories.
+//
+// The string values & category values should be wrapped in `HistoryObj`.
+//
+// If any error is occur during insert, the insert process will stop and return its error.
+func InsertMultiple(db *lazydb.LazyDB, values ...HistoryVal) error {
+	// Prevent nil database
+	if db == nil {
+		return ErrDatabaseNil
+	}
+
+	// Loop value item with core function
+	for _, val := range values {
+		err := insertValue(db, val.Category, val.Value)
+
+		// Stop loop if any error occurs
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
